requirements.txt
crewai
crewai-tools
sentence-transformers
numpy
streamlit
pydantic
requests
